third_party/blink/web_tests/fast/events/constructors/media-query-list-event-constructor-expected.txt

This tests the constructor for the MediaQueryListEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new MediaQueryListEvent('eventType').bubbles is false
PASS new MediaQueryListEvent('eventType').cancelable is false
PASS new MediaQueryListEvent('eventType').media is ""
PASS new MediaQueryListEvent('eventType').matches is false
PASS new MediaQueryListEvent('eventType', { bubbles: true, cancelable: true }).bubbles is true
PASS new MediaQueryListEvent('eventType', { bubbles: true, cancelable: true }).cancelable is true
PASS new MediaQueryListEvent('eventType').media is ""
PASS new MediaQueryListEvent('eventType').matches is false
PASS new MediaQueryListEvent('eventType', { media: 'string' }).media is "string"
PASS new MediaQueryListEvent('eventType', { matches: true }).matches is true
PASS successfullyParsed is true

TEST COMPLETE
